Moving create and get project use cases to project
